backend/pyproject.toml
[project]
name = "estudia-seguro-backend"
version = "0.1.0"
description = "FastAPI backend for Estudia Seguro academic advisory chat"
authors = [{ name = "Estudia Seguro", email = "soporte@estudiaseguro.com" }]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-dotenv>=1.0.1",
    "pydantic[email]>=2.8.0",
    "pydantic-settings>=2.2.1",
    "passlib[bcrypt]>=1.7.4",
    "supabase>=2.4.6",
    "stripe>=9.2.0",
    "httpx>=0.27.0",
    "pyjwt>=2.8.0",
    "loguru>=0.7.2",
    "sentence-transformers>=2.2.2",
    "numpy>=1.26.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.20",
    "duckduckgo-search>=6.0.0"
]
requires-python = ">=3.10"

[project.optional-dependencies]
dev = ["ruff==0.3.7", "mypy==1.9.0"]

[tool.uvicorn]
factory = false
reload = true
host = "0.0.0.0"
port = 8000

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

